+++ scratch.js
@@ -29,42 +29,44 @@ const fromFile = require('./tests/lib/_cachedPage');
 // '''George C. Papanicolaou''' (born January 23, 1943) is a [[Greece|Greek]]-[[United States|American]] [[mathematician]] who specializes in [[applied mathematics|applied]] and [[computational mathematics]], [[partial differential equation]]s, and [[stochastic process]]es.<ref name="sufac">http://math.stanford.edu/directory/faculty.html</ref> He is currently the Robert Grimmett Professor in Mathematics at [[Stanford University]].
 // `;
 
+// {{Tracklist
+// | collapsed       =
+// | headline        = Track list
+// | extra_column    = Artist(s)
+// | total_length    = 23:14
+// | all_lyrics      = [[Ramajogayya Sastry]], except where noted
+// | music_credits   =
+// | title1          = Rama Rama
+// | extra1          = [[Sooraj Santhosh]], [[Ranina Reddy]]
+// | length1         = 4:20
+// | title2          = Jatha Kalise
+// | extra2          = Sagar, [[Suchitra]]
+// | length2         = 3:44
+// | title3          = Charuseela
+// | note3           = Co-written by [[Devi Sri Prasad]]
+// | extra3          = Yazin Nizar, Devi Sri Prasad (Rap)
+// | length3         = 4:15
+// | title4          = Srimanthuda
+// | extra4          = [[M. L. R. Karthikeyan]]
+// | length4         = 2:03
+// | title5          = Jaago
+// | extra5          = [[Raghu Dixit]], [[Rita (Indian singer)|Rita]]
+// | length5         = 4:11
+// | title6          = Dhimmathirigae
+// | extra6          = Simha, [[Geetha Madhuri]]
+// | length6         = 4:41
+// }}
+
 let wiki = `
 [[File:War.png]]
 intro bit
 ==History==
-this is the [[second]] section.
+this is the [[second]] section.<ref>[http://coolyeah.yes fun times]</ref>
 
 this is the second paragraph. woo hoo.
 ==Discography==
 {{main|lkjsf}}
 hello
-{{Tracklist
-| collapsed       =
-| headline        = Track list
-| extra_column    = Artist(s)
-| total_length    = 23:14
-| all_lyrics      = [[Ramajogayya Sastry]], except where noted
-| music_credits   =
-| title1          = Rama Rama
-| extra1          = [[Sooraj Santhosh]], [[Ranina Reddy]]
-| length1         = 4:20
-| title2          = Jatha Kalise
-| extra2          = Sagar, [[Suchitra]]
-| length2         = 3:44
-| title3          = Charuseela
-| note3           = Co-written by [[Devi Sri Prasad]]
-| extra3          = Yazin Nizar, Devi Sri Prasad (Rap)
-| length3         = 4:15
-| title4          = Srimanthuda
-| extra4          = [[M. L. R. Karthikeyan]]
-| length4         = 2:03
-| title5          = Jaago
-| extra5          = [[Raghu Dixit]], [[Rita (Indian singer)|Rita]]
-| length5         = 4:11
-| title6          = Dhimmathirigae
-| extra6          = Simha, [[Geetha Madhuri]]
-| length6         = 4:41
-}}
+
 hellow world`;
-// console.log(wtf(wiki).sections(2));
+console.log(wtf(wiki).citations());